refactor: add message render component
